Moving the factory and accessor for cuda stream sync events into the detail namespace
