+++ cpp/include/raft/core/resource/cuda_event.hpp
@@ -35,30 +35,4 @@ class cuda_event_resource : public resource {
  private:
   cudaEvent_t event_;
 };
-
-/**
- * Factory that knows how to construct a
- * specific raft::resource to populate
- * the res_t.
- */
-class cuda_stream_sync_event_resource_factory : public resource_factory {
- public:
-  resource_type get_resource_type() override { return resource_type::CUDA_STREAM_SYNC_EVENT; }
-  resource* make_resource() override { return new cuda_event_resource(); }
-};
-
-/**
- * Load a cudaStream_t from a res (and populate it on the res
- * if needed).
- * @param res raft res object for managing resources
- * @return
- */
-inline cudaEvent_t& get_cuda_stream_sync_event(resources const& res)
-{
-  if (!res.has_resource_factory(resource_type::CUDA_STREAM_SYNC_EVENT)) {
-    res.add_resource_factory(std::make_shared<cuda_stream_sync_event_resource_factory>());
-  }
-  return *res.get_resource<cudaEvent_t>(resource_type::CUDA_STREAM_SYNC_EVENT);
-};
-
 }  // namespace raft::resource
